Test example code
This adds sanity tests to ensure that example code in the `examples/`
directory and `README.md` isn't outright wrong. These are not intended
to serve as unit tests for the example functionality, just as smoke
tests to catch things like API changes that require updating examples.

Including a sanity test ensures that the example code isn't outright
wrong. This is not intended to serve as unit tests for the example
functionality.

Some minor changes are included which facilitate this work:

- Use `console` instead of `bash` codeblock language

    `console` highlights `$ ` prefixed lines differently from the
    following lines, which clearly distinguishes between commands and
    input/output.

- Set `file_fixture_path`

    This allows us to use `ActiveSupport::TestCase#file_fixture`.

- Add `ReadmeTestHelper`

    This helper provides utilities for extracting code snippets from
    `README.md`.

Some of these tests revealed that either the examples were busted, or
even bugs in the implementation.

- Test README per-server configuration example

    This test revealed that the `define_` helper methods were failing to
    ensure the server supports the type of capability they were
    defining.

- Test README protocol version examples

    This revealed that the `protocol_version` accessors weren't
    available on the `Server` at all, and the examples were incorrect.

- Test README tool definition examples

    This revealed that the `define` example doesn't work, and the fix is unclear.